Explicitly cast type to int

+++ src/quiz.cpp
@@ -37,7 +37,7 @@ class Question {
 };
 
 int GetReadableScore(int score, std::size_t size) {
-  return (static_cast<double>(score) / static_cast<double>(size)) * 100;
+  return static_cast<int>((static_cast<double>(score) / static_cast<double>(size)) * 100);
 }
 
 int main() {
